+++ CHANGELOG.md
@@ -24,11 +24,12 @@ Change log
 - Keys widget now has examples built in.
 - Help menu with links to websites.
 - About dialog shows application version and primary dependency versions.
+- Support hex key codes in Key Press
 - New display system:
     - User interface shows same image as Stream Deck.
     - Text overlay on top of image, with automatic font outline.
     - Buttons image change when pressed (visual feedback).
-    - Animated gif support.
+    - Animated GIF support.
     - CPU indicator for display processing.
 
 ### Fixes